pyproject.toml
[project]
name = "prepare-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pymupdf>=1.25.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]
